Rename MetadataToASTNodeResolver to TableInfoToAST;
Using `Metadata` in the name is confusing because it's not doing
anything with the `Metadata` class

Also removes the internal `Extractor` class by lifting all the methods
to the parent class. They are private, they don't bloat the public API
surface.
