+++ server/src/main/java/io/crate/planner/node/management/ShowCreateTablePlan.java
@@ -21,8 +21,10 @@
 
 package io.crate.planner.node.management;
 
+import static io.crate.data.SentinelRow.SENTINEL;
+
 import io.crate.analyze.AnalyzedShowCreateTable;
-import io.crate.analyze.MetadataToASTNodeResolver;
+import io.crate.analyze.TableInfoToAST;
 import io.crate.data.InMemoryBatchIterator;
 import io.crate.data.Row;
 import io.crate.data.Row1;
@@ -33,8 +35,7 @@ import io.crate.planner.PlannerContext;
 import io.crate.planner.operators.SubQueryResults;
 import io.crate.sql.SqlFormatter;
 import io.crate.sql.tree.CreateTable;
-
-import static io.crate.data.SentinelRow.SENTINEL;
+import io.crate.sql.tree.Expression;
 
 public class ShowCreateTablePlan implements Plan {
 
@@ -57,7 +58,8 @@ public class ShowCreateTablePlan implements Plan {
                               SubQueryResults subQueryResults) {
         Row1 row;
         try {
-            CreateTable createTable = MetadataToASTNodeResolver.resolveCreateTable(statement.tableInfo());
+            CreateTable<Expression> createTable = new TableInfoToAST(statement.tableInfo())
+                .extractCreateTable();
             row = new Row1(SqlFormatter.formatSql(createTable));
         } catch (Throwable t) {
             consumer.accept(null, t);
