+++ server/src/test/java/io/crate/analyze/TableInfoToASTTest.java
@@ -35,12 +35,13 @@ import io.crate.metadata.Reference;
 import io.crate.metadata.doc.DocTableInfo;
 import io.crate.sql.SqlFormatter;
 import io.crate.sql.tree.CreateTable;
+import io.crate.sql.tree.Expression;
 import io.crate.test.integration.CrateDummyClusterServiceUnitTest;
 import io.crate.testing.SQLExecutor;
 import io.crate.types.ArrayType;
 import io.crate.types.DataTypes;
 
-public class MetadataToASTNodeResolverTest extends CrateDummyClusterServiceUnitTest {
+public class TableInfoToASTTest extends CrateDummyClusterServiceUnitTest {
 
     @Override
     protected boolean enableWarningsCheck() {
@@ -80,7 +81,7 @@ public class MetadataToASTNodeResolverTest extends CrateDummyClusterServiceUnitT
             .build();
         DocTableInfo tableInfo = e.resolveTableInfo("doc.test");
 
-        CreateTable node = MetadataToASTNodeResolver.resolveCreateTable(tableInfo);
+        CreateTable<Expression> node = TableInfoToAST.toCreateTable(tableInfo);
         assertEquals("CREATE TABLE IF NOT EXISTS \"doc\".\"test\" (\n" +
                      "   \"bools\" BOOLEAN,\n" +
                      "   \"bytes\" BYTE,\n" +
@@ -147,7 +148,7 @@ public class MetadataToASTNodeResolverTest extends CrateDummyClusterServiceUnitT
             .build();
         DocTableInfo tableInfo = e.resolveTableInfo("myschema.test");
 
-        CreateTable node = MetadataToASTNodeResolver.resolveCreateTable(tableInfo);
+        CreateTable<Expression> node = TableInfoToAST.toCreateTable(tableInfo);
         assertEquals("CREATE TABLE IF NOT EXISTS \"myschema\".\"test\" (\n" +
                      "   \"pk_col_one\" BIGINT NOT NULL,\n" +
                      "   \"pk_col_two\" BIGINT NOT NULL,\n" +
@@ -197,7 +198,7 @@ public class MetadataToASTNodeResolverTest extends CrateDummyClusterServiceUnitT
             .build();
         DocTableInfo tableInfo = e.resolveTableInfo("myschema.test");
 
-        CreateTable node = MetadataToASTNodeResolver.resolveCreateTable(tableInfo);
+        CreateTable<Expression> node = TableInfoToAST.toCreateTable(tableInfo);
         assertEquals("CREATE TABLE IF NOT EXISTS \"myschema\".\"test\" (\n" +
                      "   \"col_a\" TEXT NOT NULL,\n" +
                      "   \"col_b\" TEXT NOT NULL INDEX USING FULLTEXT WITH (\n" +
@@ -247,7 +248,7 @@ public class MetadataToASTNodeResolverTest extends CrateDummyClusterServiceUnitT
             .build();
         DocTableInfo tableInfo = e.resolveTableInfo("doc.test");
 
-        CreateTable node = MetadataToASTNodeResolver.resolveCreateTable(tableInfo);
+        CreateTable<Expression> node = TableInfoToAST.toCreateTable(tableInfo);
         assertEquals("CREATE TABLE IF NOT EXISTS \"doc\".\"test\" (\n" +
                      "   \"floats\" REAL,\n" +
                      "   \"shorts\" SMALLINT,\n" +
@@ -297,7 +298,7 @@ public class MetadataToASTNodeResolverTest extends CrateDummyClusterServiceUnitT
             .build();
         DocTableInfo tableInfo = e.resolveTableInfo("myschema.test");
 
-        CreateTable node = MetadataToASTNodeResolver.resolveCreateTable(tableInfo);
+        CreateTable<Expression> node = TableInfoToAST.toCreateTable(tableInfo);
         assertEquals("CREATE TABLE IF NOT EXISTS \"myschema\".\"test\" (\n" +
                      "   \"id\" BIGINT,\n" +
                      "   \"partition_column\" TEXT,\n" +
@@ -359,7 +360,7 @@ public class MetadataToASTNodeResolverTest extends CrateDummyClusterServiceUnitT
             .build();
         DocTableInfo tableInfo = e.resolveTableInfo("myschema.test");
 
-        CreateTable node = MetadataToASTNodeResolver.resolveCreateTable(tableInfo);
+        CreateTable<Expression> node = TableInfoToAST.toCreateTable(tableInfo);
         assertEquals("CREATE TABLE IF NOT EXISTS \"myschema\".\"test\" (\n" +
                      "   \"id\" BIGINT,\n" +
                      "   \"col_a\" TEXT,\n" +
@@ -422,7 +423,7 @@ public class MetadataToASTNodeResolverTest extends CrateDummyClusterServiceUnitT
             .build();
         DocTableInfo tableInfo = e.resolveTableInfo("myschema.test");
 
-        CreateTable node = MetadataToASTNodeResolver.resolveCreateTable(tableInfo);
+        CreateTable<Expression> node = TableInfoToAST.toCreateTable(tableInfo);
         assertEquals("CREATE TABLE IF NOT EXISTS \"myschema\".\"test\" (\n" +
                      "   \"s\" TEXT STORAGE WITH (\n" +
                      "      columnstore = false\n" +
@@ -468,7 +469,7 @@ public class MetadataToASTNodeResolverTest extends CrateDummyClusterServiceUnitT
                       ")")
             .build();
         DocTableInfo tableInfo = e.resolveTableInfo("test");
-        CreateTable<?> node = MetadataToASTNodeResolver.resolveCreateTable(tableInfo);
+        CreateTable<?> node = TableInfoToAST.toCreateTable(tableInfo);
         assertEquals("CREATE TABLE IF NOT EXISTS \"doc\".\"test\" (\n" +
                      "   \"col1\" TEXT,\n" +
                      "   \"col2\" INTEGER DEFAULT 2,\n" +
@@ -508,7 +509,7 @@ public class MetadataToASTNodeResolverTest extends CrateDummyClusterServiceUnitT
             .addTable("create table tbl (name varchar(10))")
             .build();
         DocTableInfo table = e.resolveTableInfo("tbl");
-        CreateTable<?> node = MetadataToASTNodeResolver.resolveCreateTable(table);
+        CreateTable<?> node = TableInfoToAST.toCreateTable(table);
         assertThat(SqlFormatter.formatSql(node), Matchers.containsString("\"name\" VARCHAR(10)"));
     }
 
@@ -518,7 +519,7 @@ public class MetadataToASTNodeResolverTest extends CrateDummyClusterServiceUnitT
             .addTable("create table tbl (xs bit(8))")
             .build();
         DocTableInfo table = e.resolveTableInfo("tbl");
-        CreateTable<?> node = MetadataToASTNodeResolver.resolveCreateTable(table);
+        CreateTable<?> node = TableInfoToAST.toCreateTable(table);
         assertThat(SqlFormatter.formatSql(node), Matchers.containsString("\"xs\" BIT(8)"));
     }
 
@@ -528,7 +529,7 @@ public class MetadataToASTNodeResolverTest extends CrateDummyClusterServiceUnitT
             .addTable("create table t (g geo_shape generated always as 'POLYGON (( 5 5, 30 5, 30 30, 5 30, 5 5 ))')")
             .build();
         DocTableInfo table = e.resolveTableInfo("t");
-        CreateTable<?> node = MetadataToASTNodeResolver.resolveCreateTable(table);
+        CreateTable<?> node = TableInfoToAST.toCreateTable(table);
         assertThat(
             SqlFormatter.formatSql(node),
             Matchers.containsString("\"g\" GEO_SHAPE GENERATED ALWAYS AS 'POLYGON (( 5 5, 30 5, 30 30, 5 30, 5 5 ))")
@@ -541,7 +542,7 @@ public class MetadataToASTNodeResolverTest extends CrateDummyClusterServiceUnitT
             .addTable("create table t (geo_arr array(geo_shape) INDEX using QUADTREE with (precision='1m', distance_error_pct='0.25'))")
             .build();
         DocTableInfo table = e.resolveTableInfo("t");
-        CreateTable<?> node = MetadataToASTNodeResolver.resolveCreateTable(table);
+        CreateTable<?> node = TableInfoToAST.toCreateTable(table);
         assertThat(
             SqlFormatter.formatSql(node),
             Matchers.containsString("""
